requirements.txt
asgiref==3.8.1
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
